requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
transformers
torch
python-dotenv
requests
accelerate